fix(dagent): use none network mode & add unit tests
