+++ golang/pkg/builder/container/container_builder.go
@@ -297,32 +297,23 @@ func (dc *DockerContainerBuilder) WithPostStartHooks(hooks ...LifecycleFunc) Bui
 	return dc
 }
 
-// Creates the container using the configuration given by 'With...' functions.
-func (dc *DockerContainerBuilder) Create() (Container, error) {
-	dc.logWrite(fmt.Sprintf("Creating container: %s", util.Fallback(dc.containerName, dc.imageWithTag)))
-	if err := dc.prepareImage(); err != nil {
-		dc.logWrite(fmt.Sprintf("Failed to prepare image: %s", err.Error()))
-		return nil, err
-	}
-
-	if dc.withoutConflict {
-		err := dockerHelper.DeleteContainerByName(dc.ctx, dc.client, dc.containerName)
-		if err != nil {
-			dc.logWrite(fmt.Sprintf("Failed to resolve conflict during creating the container: %v", err))
-			return nil, err
-		}
+func builderToDockerConfig(dc *DockerContainerBuilder) (hostConfig *container.HostConfig, containerConfig *container.Config, err error) {
+	hostConfig = &container.HostConfig{}
+	containerConfig = &container.Config{}
+	if dc.containerName != "" {
+		containerConfig.Hostname = dc.containerName
 	}
 
 	portListNat := portListToNatBinding(dc.portRanges, dc.portList)
 	exposedPortSet := getPortSet(dc.portList)
-	hostConfig := &container.HostConfig{
+	hostConfig = &container.HostConfig{
 		Mounts:       dc.mountList,
 		PortBindings: portListNat,
 		AutoRemove:   dc.remove,
 		ExtraHosts:   dc.extraHosts,
 	}
 
-	containerConfig := &container.Config{
+	containerConfig = &container.Config{
 		Image:        dc.imageWithTag,
 		Tty:          dc.tty,
 		Env:          dc.envList,
@@ -348,29 +339,47 @@ func (dc *DockerContainerBuilder) Create() (Container, error) {
 	if dc.networkMode != "" {
 		dc.logWrite(fmt.Sprintf("Provided networkMode: %s", dc.networkMode))
 	}
-	if nw := container.NetworkMode(dc.networkMode); !nw.IsPrivate() {
+	if nw := container.NetworkMode(dc.networkMode); !nw.IsPrivate() || nw.IsNone() {
 		hostConfig.NetworkMode = nw
 	} else {
 		networkCreateResult, err := createNetworks(dc)
 		if err != nil {
-			return nil, errors.Join(err, fmt.Errorf("error creating enlisted containers networks: %v", dc.networks))
+			return nil, nil, errors.Join(err, fmt.Errorf("error creating enlisted containers networks: %v", dc.networks))
 		}
 
 		dc.networkMap = networkCreateResult
 		dc.logWrite(fmt.Sprintf("Container network: %v", maps.Keys(dc.networkMap)))
 	}
 
-	var name string
-	if dc.containerName != "" {
-		name = dc.containerName
-		containerConfig.Hostname = dc.containerName
+	return hostConfig, containerConfig, nil
+}
+
+// Creates the container using the configuration given by 'With...' functions.
+func (dc *DockerContainerBuilder) Create() (Container, error) {
+	dc.logWrite(fmt.Sprintf("Creating container: %s", util.Fallback(dc.containerName, dc.imageWithTag)))
+	if err := dc.prepareImage(); err != nil {
+		dc.logWrite(fmt.Sprintf("Failed to prepare image: %s", err.Error()))
+		return nil, err
+	}
+
+	if dc.withoutConflict {
+		err := dockerHelper.DeleteContainerByName(dc.ctx, dc.client, dc.containerName)
+		if err != nil {
+			dc.logWrite(fmt.Sprintf("Failed to resolve conflict during creating the container: %v", err))
+			return nil, err
+		}
 	}
 
 	if hookError := execHooks(dc, nil, dc.hooksPreCreate); hookError != nil {
 		dc.logWrite(fmt.Sprintln("Container pre-create hook error: ", hookError))
 	}
 
-	containerCreateResp, err := dc.client.ContainerCreate(dc.ctx, containerConfig, hostConfig, nil, nil, name)
+	hostConfig, containerConfig, err := builderToDockerConfig(dc)
+	if err != nil {
+		return nil, err
+	}
+
+	containerCreateResp, err := dc.client.ContainerCreate(dc.ctx, containerConfig, hostConfig, nil, nil, dc.containerName)
 	if err != nil {
 		dc.logWrite(fmt.Sprintln("Container create failed: ", err))
 	}
@@ -524,7 +533,10 @@ func execHooks(dc *DockerContainerBuilder, cont *types.Container, hooks []Lifecy
 }
 
 func portListToNatBinding(portRanges []PortRangeBinding, portList []PortBinding) map[nat.Port][]nat.PortBinding {
-	portMap := make(map[nat.Port][]nat.PortBinding)
+	if len(portRanges) == 0 && len(portList) == 0 {
+		return nil
+	}
+	portMap := map[nat.Port][]nat.PortBinding{}
 
 	for _, p := range portRanges {
 		// example complete portSpec
@@ -563,7 +575,10 @@ func portListToNatBinding(portRanges []PortRangeBinding, portList []PortBinding)
 }
 
 func getPortSet(portList []PortBinding) nat.PortSet {
-	portSet := make(nat.PortSet)
+	if len(portList) == 0 {
+		return nil
+	}
+	portSet := nat.PortSet{}
 
 	for _, port := range portList {
 		exposedPort, _ := nat.NewPort("tcp", fmt.Sprint(port.ExposedPort))
