+++ golang/pkg/builder/container/container_builder_export_test.go
@@ -1,0 +1,3 @@
+package container
+
+var BuilderToDockerConfig = builderToDockerConfig
