--- golang/pkg/builder/container/docker_container_test.go
@@ -1,13 +1,0 @@
-//go:build integration
-// +build integration
-
-package container_test
-
-import (
-	"os"
-	"testing"
-)
-
-func TestMain(m *testing.M) {
-	os.Exit(m.Run())
-}
